remove warnings
svn path=/trunk/kdenlive/; revision=1257

+++ kdenlive/kmmscreen.cpp
@@ -190,9 +190,9 @@ namespace Gui {
 
 //virtual
     void KMMScreen::wheelEvent(QWheelEvent * e) {
-	int factor = 1;
+	double factor = 1.0;
 	if (( e->state() & ControlButton) == ControlButton) factor = m_app->getDocument()->framesPerSecond();
-	GenTime newSeek = seekPosition() - GenTime(e->delta() / 120 * factor, m_app->getDocument()->framesPerSecond());
+	GenTime newSeek = seekPosition() - GenTime((int) e->delta() / 120 * factor, m_app->getDocument()->framesPerSecond());
 	if (newSeek < GenTime())
 	    newSeek = GenTime(0);
 	if (newSeek > m_clipLength)
